Enhance UI alignment on the homepage and reorganize CSS files
