+++ app/login/layout.jsx
@@ -1,4 +1,4 @@
-import ToggleThemeButton from "../../components/toggleTheme/ToggleTheme";
+import ToggleThemeButton from "../../components/UI/ToggleTheme";
 
 export default function LoginLayout({ children }) {
   return (
