+++ components/UI/ToggleTheme.jsx
@@ -2,7 +2,6 @@
 
 import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import "./ToggleTheme.css";
 
 export default function ToggleThemeButton() {
   const { t } = useTranslation();
