+++ components/search/Search.css
@@ -1,5 +1,4 @@
 .search-container {
-  margin-bottom: 20px;
   text-align: center;
 }
