+++ components/sort/Sort.css
@@ -1,0 +1,3 @@
+.sort-button {
+  margin-top: 20px;
+}
