--- components/toggleTheme/ToggleTheme.css
@@ -1,57 +1,0 @@
-.light-theme {
-  background: linear-gradient(135deg, white 0%, purple 100%);
-  color: black;
-}
-
-.dark-theme {
-  background: linear-gradient(135deg, #2c3e50 0%, purple 100%);
-  color: white;
-}
-
-.dark-theme h1 {
-  color: white;
-}
-
-.theme-switcher {
-  position: relative;
-  display: inline-block;
-  margin-right: 10px;
-}
-
-.theme-switcher button.active {
-  background-color: #e0e0e0;
-  color: #000;
-}
-
-.theme-switcher button:not(:last-child) {
-  border-bottom: none;
-}
-
-.theme-switcher-content {
-  display: none;
-  position: absolute;
-  left: 0;
-  background-color: #f9f9f9;
-  min-width: 160px;
-  z-index: 1;
-  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
-  border-radius: 5px;
-  overflow: hidden;
-  padding-right: 20px;
-  margin-bottom: 20px;
-}
-
-.theme-switcher:hover .theme-switcher-content {
-  display: block;
-}
-
-.theme-switcher-content button {
-  display: block;
-  width: 100%;
-  text-align: left;
-}
-
-.system-mode-button {
-  display: flex;
-  align-items: center;
-}
